ref: Refactor project structure, change file names to conventional
