+++ main.go
@@ -10,8 +10,8 @@ import (
 	"github.com/google/uuid"
 
 	"github.com/Yulian302/qugopy/config"
+	"github.com/Yulian302/qugopy/internal/queue"
 	"github.com/Yulian302/qugopy/models"
-	"github.com/Yulian302/qugopy/queue"
 	"github.com/gin-gonic/gin"
 	"github.com/go-redis/redis"
 )
